Fixed int overflow bug

+++ src/main/java/com/cobalt/jira/plugin/epic/data/DataManager.java
@@ -27,7 +27,7 @@ import java.util.*;
 public class DataManager {
     //issues that has changed in the given time span excluding a list of given issues
     private static final String QUERY = "(status CHANGED FROM (%s) AFTER %s OR status CHANGED TO (%s) AFTER %s) AND issuetype not in (%s) ORDER BY updated DESC";
-    private static final int MAX_DAYS = 28;
+    private static final long MAX_DAYS = 28;
     private static final String DEFAULT_QUERY = String.format(QUERY, StatusUtil.getInitialStates(), "-" + MAX_DAYS + "d", StatusUtil.getEndStates(), "-" + MAX_DAYS + "d", "Epic");
 
     private NaryTree tree;
